backend/app/utils/security.py
from datetime import datetime, timedelta, timezone

from jose import jwt

from ..config import get_settings

settings = get_settings()


def create_access_token(subject: str, role: str) -> str:
    expires_at = datetime.now(timezone.utc) + timedelta(minutes=settings.access_token_expire_minutes)
    payload = {
        "sub": subject,
        "role": role,
        "exp": expires_at,
    }
    return jwt.encode(payload, settings.jwt_secret_key, algorithm="HS256")
